refactored dt test and moved to wrapper_dt_wasm
