+++ rust/app/Cargo.toml
@@ -14,5 +14,4 @@ serde_json = "1.0.85"
 async-trait = "0.1.58"
 by_address = "1.1.0"
 lazy_static = "1.4.0"
-tokio = { version = "1.22.0", features = ["rt", "macros"] }
 futures = "0.3.25"
