--- rust/app/tests/dt/dt_example/src/polywrap/resources/external_resource.rs
@@ -1,3 +1,0 @@
-pub enum ExternalResource {
-    Log = 1337,
-}
