--- rust/app/tests/dt/dt_example/src/polywrap/resources/internal_resource.rs
@@ -1,3 +1,0 @@
-pub enum InternalResource {
-    Log = 1338
-}
