--- rust/app/tests/dt/dt_example/src/polywrap/resources/mod.rs
@@ -1,4 +1,0 @@
-pub mod external_resource;
-pub mod internal_resource;
-pub use external_resource::ExternalResource;
-pub use internal_resource::InternalResource;
