+++ rust/reim_wrapper_dt_wasm/Cargo.toml
@@ -7,3 +7,7 @@ edition = "2021"
 reim_dt = { path = "../reim_dt" }
 futures = "0.3.25"
 async-trait = "0.1.58"
+[dev_dependencies]
+tokio = { version = "1.22.0", features = ["rt", "macros"] }
+reim_host_dt = { path = "../reim_host_dt" }
+reim_host_dt_wasm = { path = "../reim_host_dt_wasm" }
