+++ rust/test_cases/wrappers/dt_example/package.json
@@ -7,8 +7,7 @@
         "build:wasm": "RUSTFLAGS=\"-C link-arg=-z -C link-arg=stack-size=65536 -C link-arg=--import-memory\" cargo build --target wasm32-unknown-unknown --release --target-dir ./target",
         "build:bindgen": "wasm-bindgen ./target/wasm32-unknown-unknown/release/module.wasm --out-dir ./build --out-name bg_module.wasm",
         "build:snip": "wasm-snip ./build/bg_module.wasm -o ./build/snipped_module.wasm",
-        "build:opt": "wasm-opt --asyncify -Os ./build/snipped_module.wasm -o ./build/wrap.wasm",
-        "postbuild": "reim_cli manifest"
+        "build:opt": "wasm-opt --asyncify -Os ./build/snipped_module.wasm -o ./build/wrap.wasm"
     },
     "dependencies": {
     },
