+++ rust/test_cases/wrappers/dt_example/schema.graphql
