+++ rust/test_cases/wrappers/dt_example/src/polywrap/mod.rs
@@ -1,4 +1,3 @@
-pub mod resources;
 pub mod exports;
 pub mod internal_wrap_module;
 pub use exports::_dt_receive;
